Remove endpoint
