+++ pkg/server/conn.go
@@ -1,12 +1,14 @@
 package server
 
 import (
+	"fmt"
 	"net"
 	"os"
 	"sync"
 	"time"
 
 	"github.com/lucheng0127/virtuallan/pkg/packet"
+	"github.com/lucheng0127/virtuallan/pkg/users"
 	"github.com/lucheng0127/virtuallan/pkg/utils"
 	log "github.com/sirupsen/logrus"
 	"github.com/songgao/water"
@@ -19,8 +21,10 @@ type UClient struct {
 	NetToIface chan *packet.VLPkt
 	Once       sync.Once
 	User       string
-	IP         string
+	IP         net.IP
 	Login      string
+	Beat       chan string
+	Svc        *Server
 }
 
 var UPool map[string]*UClient
@@ -40,7 +44,22 @@ func (client *UClient) Close() {
 		log.Error(err)
 	}
 
+	client.Svc.ReleaseIP(client.IP)
+	delete(users.UserEPMap, client.User)
 	delete(UPool, client.RAddr.String())
+}
+
+func (client *UClient) Countdown() {
+	for {
+		select {
+		case <-client.Beat:
+			continue
+		case <-time.After(50 * time.Second):
+			log.Infof("endpoint %s with raddr %s don't get keepalive pkt for long time, close it\n", client.IP, client.RAddr.String())
+			client.Close()
+			return
+		}
+	}
 }
 
 func (client *UClient) Handle() {
@@ -123,6 +142,116 @@ func (svc *Server) CreateClientForAddr(addr *net.UDPAddr, conn *net.UDPConn) (*U
 	client.NetToIface = make(chan *packet.VLPkt, 1024)
 	client.Login = time.Now().Format("2006-01-02 15:04:05")
 	client.Once = sync.Once{}
+	client.Beat = make(chan string)
+	client.Svc = svc
+
 	UPool[addr.String()] = client
+
 	return client, nil
+}
+
+func (svc *Server) ListenAndServe() error {
+	if !utils.ValidatePort(svc.Port) {
+		return fmt.Errorf("invalidate port %d", svc.Port)
+	}
+
+	addr, err := net.ResolveUDPAddr("udp4", fmt.Sprintf("0.0.0.0:%d", svc.Port))
+	if err != nil {
+		return err
+	}
+
+	ln, err := net.ListenUDP("udp4", addr)
+	if err != nil {
+		return err
+	}
+	defer ln.Close()
+
+	for {
+		// Max vlpkt len 1502 = 1500(max ethernet pkt) + 2(vlheader)
+		// for encrypted data len should be n*16(aes block size) + 16(key len)
+		// so buf len should be 94 * 16 + 16 = 1520
+		var buf [65535]byte
+		n, addr, err := ln.ReadFromUDP(buf[:])
+		if err != nil {
+			return err
+		}
+
+		if n < 2 {
+			continue
+		}
+
+		// For wrong AES key, will return pkt to nill or unsupported pkt error, just skip
+		pkt, err := packet.Decode(buf[:n])
+		if pkt == nil {
+			continue
+		}
+
+		if err != nil {
+			if utils.IsUnsupportedPkt(err) {
+				log.Warn(err)
+				continue
+			}
+
+			log.Error("parse packet ", err)
+		}
+
+		// TODO(shawnlu): Add close conn
+		switch pkt.Type {
+		case packet.P_AUTH:
+			u, p := pkt.VLBody.(*packet.AuthBody).Parse()
+
+			// Check user logged
+			if _, ok := users.UserEPMap[u]; ok {
+				svc.SendResponse(ln, packet.RSP_USER_LOGGED, addr)
+				continue
+			}
+
+			// Auth user
+			err = users.ValidateUser(svc.userDb, u, p)
+
+			if err != nil {
+				log.Warn(err)
+				svc.SendResponse(ln, packet.RSP_AUTH_REQUIRED, addr)
+				continue
+			}
+
+			users.UserEPMap[u] = addr.String()
+			log.Infof("client %s login to %s succeed\n", addr.String(), u)
+
+			// Create client for authed addr
+			client, err := svc.CreateClientForAddr(addr, ln)
+			if err != nil {
+				log.Errorf("create authed client %s\n", err.Error())
+			}
+			client.User = u
+
+			log.Infof("client %s auth succeed", addr.String())
+		case packet.P_KEEPALIVE:
+			// Handle keepalive
+			err = HandleKeepalive(pkt.VLBody.(*packet.KeepaliveBody).Parse(), addr.String(), svc)
+
+			if err != nil {
+				if utils.IsUnauthedErr(err) {
+					continue
+				}
+
+				svc.SendResponse(ln, packet.RSP_IP_CONFLICET, addr)
+				log.Warnf("heartbeat from %s %s, send ip conflicet response", addr.String(), err.Error())
+			}
+		case packet.P_RAW:
+			// Get authed client from UPool
+			client, ok := UPool[addr.String()]
+			if !ok {
+				svc.SendResponse(ln, packet.RSP_AUTH_REQUIRED, addr)
+				continue
+			}
+
+			go client.HandleOnce()
+
+			client.NetToIface <- pkt
+		default:
+			log.Debug("unknow stream, do nothing")
+			continue
+		}
+	}
 }
