+++ pkg/server/server.go
@@ -10,8 +10,6 @@ import (
 	"sync"
 
 	"github.com/lucheng0127/virtuallan/pkg/config"
-	"github.com/lucheng0127/virtuallan/pkg/packet"
-	"github.com/lucheng0127/virtuallan/pkg/users"
 	"github.com/lucheng0127/virtuallan/pkg/utils"
 	log "github.com/sirupsen/logrus"
 	"github.com/urfave/cli/v2"
@@ -26,6 +24,15 @@ type Server struct {
 	IPStart net.IP
 	IPCount int
 	MLock   sync.Mutex
+}
+
+func NewServer() *Server {
+	svc := new(Server)
+
+	svc.UsedIP = make([]int, 0)
+	svc.MLock = sync.Mutex{}
+
+	return svc
 }
 
 func (svc *Server) SetupLan() error {
@@ -48,112 +55,6 @@ func (svc *Server) SetupLan() error {
 	return nil
 }
 
-func (svc *Server) ListenAndServe() error {
-	if !utils.ValidatePort(svc.Port) {
-		return fmt.Errorf("invalidate port %d", svc.Port)
-	}
-
-	addr, err := net.ResolveUDPAddr("udp4", fmt.Sprintf("0.0.0.0:%d", svc.Port))
-	if err != nil {
-		return err
-	}
-
-	ln, err := net.ListenUDP("udp4", addr)
-	if err != nil {
-		return err
-	}
-	defer ln.Close()
-
-	for {
-		// Max vlpkt len 1502 = 1500(max ethernet pkt) + 2(vlheader)
-		// for encrypted data len should be n*16(aes block size) + 16(key len)
-		// so buf len should be 94 * 16 + 16 = 1520
-		var buf [65535]byte
-		n, addr, err := ln.ReadFromUDP(buf[:])
-		if err != nil {
-			return err
-		}
-
-		if n < 2 {
-			continue
-		}
-
-		// For wrong AES key, will return pkt to nill or unsupported pkt error, just skip
-		pkt, err := packet.Decode(buf[:n])
-		if pkt == nil {
-			continue
-		}
-
-		if err != nil {
-			if utils.IsUnsupportedPkt(err) {
-				log.Warn(err)
-				continue
-			}
-
-			log.Error("parse packet ", err)
-		}
-
-		// TODO(shawnlu): Add close conn
-		switch pkt.Type {
-		case packet.P_AUTH:
-			u, p := pkt.VLBody.(*packet.AuthBody).Parse()
-
-			// Check user logged
-			if _, ok := users.UserEPMap[u]; ok {
-				svc.SendResponse(ln, packet.RSP_USER_LOGGED, addr)
-				continue
-			}
-
-			// Auth user
-			err = users.ValidateUser(svc.userDb, u, p)
-
-			if err != nil {
-				log.Warn(err)
-				svc.SendResponse(ln, packet.RSP_AUTH_REQUIRED, addr)
-				continue
-			}
-
-			users.UserEPMap[u] = addr.String()
-			log.Infof("client %s login to %s succeed\n", addr.String(), u)
-
-			// Create client for authed addr
-			client, err := svc.CreateClientForAddr(addr, ln)
-			if err != nil {
-				log.Errorf("create authed client %s\n", err.Error())
-			}
-			client.User = u
-
-			log.Infof("client %s auth succeed", addr.String())
-		case packet.P_KEEPALIVE:
-			// Handle keepalive
-			err = HandleKeepalive(pkt.VLBody.(*packet.KeepaliveBody).Parse(), addr.String())
-
-			if err != nil {
-				if utils.IsUnauthedErr(err) {
-					continue
-				}
-
-				svc.SendResponse(ln, packet.RSP_IP_CONFLICET, addr)
-				log.Warnf("heartbeat from %s %s, send ip conflicet response", addr.String(), err.Error())
-			}
-		case packet.P_RAW:
-			// Get authed client from UPool
-			client, ok := UPool[addr.String()]
-			if !ok {
-				svc.SendResponse(ln, packet.RSP_AUTH_REQUIRED, addr)
-				continue
-			}
-
-			go client.HandleOnce()
-
-			client.NetToIface <- pkt
-		default:
-			log.Debug("unknow stream, do nothing")
-			continue
-		}
-	}
-}
-
 func (svc *Server) Teardown() {
 	err := utils.DelLinkByName(svc.Bridge)
 	if err != nil {
@@ -169,12 +70,9 @@ func (svc *Server) HandleSignal(sigChan chan os.Signal) {
 	svc.Teardown()
 }
 
-// TODO(shawnlu): Add dhcp
 func Run(cCtx *cli.Context) error {
 	// New server and do cfg parse
-	svc := new(Server)
-	svc.UsedIP = make([]int, 0)
-	svc.MLock = sync.Mutex{}
+	svc := NewServer()
 
 	cfgDir := cCtx.String("config-dir")
 	cfg, err := config.LoadConfigFile(config.GetCfgPath(cfgDir))
@@ -206,7 +104,10 @@ func Run(cCtx *cli.Context) error {
 
 	// Run web server
 	if svc.ServerConfig.WebConfig.Enable {
-		webSvc := &webServe{port: svc.ServerConfig.WebConfig.Port}
+		webSvc := &webServe{
+			port: svc.ServerConfig.WebConfig.Port,
+			svc:  svc,
+		}
 		go webSvc.Serve()
 		log.Info("run web server on port ", webSvc.port)
 	}
