+++ pkg/server/web.go
@@ -11,6 +11,7 @@ import (
 
 type webServe struct {
 	port int
+	svc  *Server
 }
 
 type EpEntry struct {
@@ -35,7 +36,7 @@ func listEpEntries(c *gin.Context) {
 			User:  user,
 			Addr:  addr,
 			Iface: c.Iface.Name(),
-			IP:    c.IP,
+			IP:    c.IP.String(),
 			Login: c.Login,
 		})
 	}
